+++ polygonizer3d.cpp
@@ -343,7 +343,7 @@ Polygonizer3d::~Polygonizer3d()
 void Polygonizer3d::setInsideCellIndices(std::vector<GridIndex> indices) {
     _insideIndices.clear();
     _insideIndices.reserve(indices.size());
-    for (int i = 0; i < (int)indices.size(); i++) {
+    for (unsigned int i = 0; i < indices.size(); i++) {
         _insideIndices.push_back(indices[i]);
     }
 }
@@ -392,7 +392,7 @@ double Polygonizer3d::_getVertexFieldValue(GridIndex g) {
         assert(!_isScalarFieldSet);
 
         glm::vec3 p = _getVertexPosition(g);
-        _vertexValues.set(g, _field->getFieldValue(p));
+        _vertexValues.set(g, (float)_field->getFieldValue(p));
     }
 
     return _vertexValues(g);;
@@ -480,7 +480,7 @@ std::vector<GridIndex> Polygonizer3d::_findSurfaceCells() {
     std::vector<GridIndex> surfaceCells;
     _isCellDone.fill(false);
 
-    for (int i = 0; i < (int)_insideIndices.size(); i++) {
+    for (unsigned int i = 0; i < _insideIndices.size(); i++) {
         GridIndex cell = _insideIndices[i];
         
         if (_isCellDone(cell)) {
@@ -701,7 +701,7 @@ void Polygonizer3d::_polygonizeCell(GridIndex g, double isolevel, EdgeGrid &edge
 void Polygonizer3d::_calculateSurfaceTriangles() {
     _surface.clear();
     EdgeGrid edges(_isize, _jsize, _ksize);
-    for (int i = 0; i < (int)_surfaceCells.size(); i++) {
+    for (unsigned int i = 0; i < _surfaceCells.size(); i++) {
         _polygonizeCell(_surfaceCells[i], _surfaceThreshold, edges);
     }
 }
